Setup the initial module hierarchy
So that people would know where to find and put code. I also started
working on an implementation of the `ls` command. It still needs some
work to clean things up and test things and it still needs the
identification of state for each patch implemented. But it is a good
start.

ps-id: CF297504-7A41-490C-AA3E-5180B6C00EF1