Make ls subcommand repo relative to cwd
Previously it was hard coded from the local development environment.
However, now I have switch the repository discover to start at the
current working directory and search from their up the file system to
find a git directory.

I also commented out some incomplete functionality in here so that the
command would be in a more usable state. I also updated the status in
the README to appropriately represent the current state of this command.

ps-id: FBC67068-F8BE-4706-805E-FD9711531D30