Extract some things to ps::git module
This involved first formalizing that the ps::git module is to be an
extension of libgit2 and not a wrapper. Beyond that we standardized on
a GitError type that will also wrap the libgit2 error type as well as
introduce our own higher level errors.

So we updated the existing functions to use the newly standardized error
type.

We also added the create_cwd_repo() and branch_upstream_name() functions
to help streamline things a little bit and cleaned up existing
functionality where these could be used instead.

ps-id: 27708F46-4F24-42AB-95F9-3C3EEB51AAD4