Got min ver of cherry_pick_no_working_copy working
This was just an exploration to see if I could get a cherry pick
implemented that wouldn't involve messing with the local state by
checking out and cherry picking, etc.

The answer seems to be yes at this point as it seems to be working.

Next need to do the following:
- figure out what happens when there is a conflict and make sure it has
  handling
- address all the unwrap() calls
- implement a cherry_pick range of commits

This is all so that I can use the cherry pick range of commits to
implement the addPatchStackIdToCommit function so that we can add the
patch stack id to a commit when we need to.