Implement basics of request review
This doesn't complete request review. But it starts to lay the
foundation of the functionality. It currently is doing the adding of
patch stack ids when necessary, the creation of the branch, as well as
the cherry picking of the patch into the branch.

There is definitely room for work in terms of error handling as
everything is basically force unwrapped right now. There is also work
that needs to happen to push the branch up the remote.

ps-id: D5DDB99F-61B8-4058-878C-E2D4DBD89DBE