Add push functionality to rr command
The intention behind this change is to get the functionality in play to
push the associated rr branch up to the remote.

I am doing this because it is part of the core functionality of the rr
command.

The approach I took was to get the remote name, and build the refspecs
that would be passed to the external `git-push` command. The reason that
I have opted for shelling out to the external git command is because it
seems like handling all the different auth options inside of libgit2 is
quite involved to facilitate push. Therefore, to avoid that I am
shelling out to the external program. This has some overhead but it
isn't an operation that should be used repetedly. In the case of the rr
command it is really the last step that occurs in that process.

ps-id: ED1FB810-D2D2-4A03-B772-B47414FB8460