Add error reporting to utils::exceute()
My intention with this change is to improve the utils::execute()
function's abstraction so that I can have it return the
std::result::Result<(), ExecuteError> allowing consumers to identify
success of void or an error of ExecuteError. This should allow consumers
to handle the various different types of errors.

I am doing this to make consuming the utils::excecute() abstraction a
bit cleaner while also facilitating robust error handling when
necessary.  This is important because I have decided we are going to
implement git-ps using a hybrid approach, some things with libgit2 and
some things executing the external git binary using utils::excute().  So
utils::excute() is going to play a pretty big role overall in git-ps.

The approach I took involved evolving the utils::execute() function to
return the std::result::Result<(), ExecuteError>, defining all the error
cases in ExecuteError and mapping the failure cases to the specific
error cases in ExecuteError.  Once I had that working I was then able to
reworking all the consumption points to deal with the () successes as
well as use Debug formatting, `{:?}`, rather than Display formatting,
`{}` because the ExceuteError doesn't implement the Display trait.

ps-id: 13E270D6-0E89-4679-9033-E1277AADF947