Add ps::pull() function
I am adding the ps::pull() function to be part of the public interface.

I am doing this so that it can be part of the ps library public
interface and we can use it through git-ps.

The approach I took was defining a new PullError type to identify the
possible failure cases and have the ps:pull() -> Result<(), PullError>.