Change pull cmd to use ps::pull()
My intension with this was to replace the old pull command
implementation with the use of ps:pull().

I did this because I want to remove duplication and standardize on using
the new ps::pull() function.

My approach was just replacing the utils::execute() usage with ps:pull()
and their associated imports.