Add ps::ps::pull to lib crate public interface
My intension to expose the ps::ps::pull() function as part of the public
interface of the lib crate and use it in binary.

The reason I am doing this is because I am trying to standardize the
public interface of the lib crate and use it in the command line tool.

I first moved the pull() and PullError code from the ps/ps/mod.rs into
the ps/ps/pull.rs file. I did this to just organize the code a bit more
as the ps/ps/mod.rs was getting a bit large. Then I proceeded to add the
pub mod pull to the ps/ps/mod.rs to define the ps/ps/pull.rs module
within the ps::ps module and expose it. Then I proceeded to nuke the
old ps/commands/pull.rs implementation that I don't want to use anymore.
Then I added the pull() and PullError to the public interface of the lib
crate and utilized them in the main.rs.