Move ps/ps/pull.rs up to ps/pull.rs
My intention is to pull it up to a level where it is clearer and to see
if we can get rid of this nested ps/ps/ weirdness.

Hopefully we find this makes the codebase more easily understood in the
future.

My approach was simply moving the file and then correcting the mod and
uses of it.

This should be a completely backwards compatible change as the public
interface of the lib is exactly the same.