Reorganize code base
My intention with this change is to reorganize the code base to to
better group functionality specific to the command line tool and better
formalize code specific to the library.

I am doing this because I think having a well organized code base is
huge for making it easy for people to contribute to open source
projects.

The approach I am taking is to basically create two groupings. One
grouping is the library and what makes up the library. This is basically
`lib` and the `ps` module. Then the other grouping is the command line
tool and what makes it up. Basically the `main` module and the
`commands` module.