Make git::create_cwd_repo() have it's own error
My intention with this is to switch the git::create_cwd_repo() error
from the more generic GitError type to the new CreateCwdRepositoryError
type.

I am doing this so that in usage of this function I am more easily able
to use the ? operation and be able to handle the mapping between the
CreateCwdRepositoryError type and the consumers error types.

My approach for doing this is simply to create the new error type and to
replace the existing one. This doesn't break any existing code as all
the conusmers at the moment are just using .unwrap().

ps-id: 9591799A-FC26-458E-8B57-8A87E29E9162