Add branch plumbing command
My intention with this is to add the `branch` subcommand.

I am doing this because we decided that we thought it would be useful to
have this as a plumbing command so that people could use it in building
up their own custom scripts.

The approach I took was to first add functionality to the BranchError
for fmt::Display. This allows me to use interpolation for the error and
have it render in human readable format. Then I added the
src/commands/plumbing/branch.rs file with the branch function to hold
the implementation of the branch command.