Extract git::ext_force_push & restructure
My intention was to extract the ext_force_push func as well as replace
the implementation of rr with a version that uses the branch() func.

I did this because I am trying to cleanup and build up the foundational
components of all of the commands.

My approach was to just do the above two things but then I ran into the
confrontation of git2 concepts needing to exist in functions like
branch. So the stance I decided to take was to allow them to exist as
arguments and return values rather than trying to wrap them or
completely hide them away. Right now I am thinking this might be ok for
what I am thinking of as **plumbing** style functionality of the ps
library. The **porcelain** style (a.k.a. the public interface) probably
shouldn't expose git2 at all.