Change ext_force_push() to ext_push()
I changed ext_force_push() to ext_push() because I made it more flexible
to deal with force pushes as well as non forceful pushes.

I did this because I needed a non-forceful push to support the integrate
command in the future.

To did this I simply added the force argument to the function as a
boolean and implemented a conditional around that to change the behavior
of the function. Then I updated the current calling locations to use the
new name and additional parameter.

ps-id: DD6827E5-47AE-472F-9935-560DAB18C4AF