Add mod test to ps/plumbing module for tests
Add the test utilities module to the ps/plumbing module only for tests
cases. That way I can use the test utilities in tests but not in normal
builds.

I want to do this so that the test suite can actually be useful.

The approach I have taken is to simply conditionally declare the test
module as part of the ps/plumbing module and expose it publicly.

ps-id: 7DCC2776-2407-43DC-83AE-0DD02F034737