Add patch state management operations
I am adding the patch state management operations so that I can use them
to properly track the association between branch (or other state in the
future) and patch.

This will allow us to use the stored state to simplify the user
interface so that users don't have think about or know about the
branches that are used in the background.

The approach I took was to simply provide a function to read the stored
state, a function to write the in memory state to disk, and then another
function to handle the process of reading the state from disk, adding an
item to it, and then writing it back to disk.

ps-id: ABF99449-C8C8-4388-BC37-1611103F71EA