Add patch state persistence to rr command
I am adding state persistence to the rr command so that I can thes use
that state in the branch command to know if I need to create a new
branch for the patch or if one has already been created for it. Also I
am going to use the state in the ls command to let the user know if a
patch has been pushed to the remote or if the patch has been already
been requested for review.

To do this required making some changes.

- had to modify the branch() to return not only the libgit2 Branch but
  also the associated patch_id
- changed the extract_ps_id() function to return a Uuid so that
  throughout the entire app patch ids are the Uuid type
- decided to change the possible states down to the 3, PushedToRemote,
  RequestedReview, and Published as those are the only states that make
  sense.

Once the above were done I was able to actually add some logic to
persist the patch state when an rr command occurs.

ps-id: 38B0AEAF-3140-4CCD-87C3-2D4D4A819467