Add patch meta data checking & creation to branch()
I am adding patch meta data checking & creation to the branch function
so that when a patch is first associated to a branch, that state is
recorded. This allows us to also read the state, check if there is
already a branch associated with the patch, if so use that branch name,
if not then generate a new one.

This is important because it prevents issues where a user might change
the summary of a patch after initially doing a branch and it resulting
in generation of a totally different branch but for the same patch.

I initially started out collapsing things into a fetch_patch_meta_data()
function but then after using it realized that it isn't a great fit for
branch() because I need to not only get the meta data but also update
and write the meta data once the branch creation and cherry-picking is
complete.