Modify branch() to record patch meta data if none
Modify the branch() function to only record patch meta data if there was
no patch meta data for the specified patch id.

This prevents us from accidentally overwriting future patch states. For
example lets say we already requested review for a patch. We wouldn't
want that to be overwritten with a state of BranchCreated.

My approach for this was simply wrapping the updating and persistence of
the patch meta data in a conditional checking if patch meta data
previously existed for this patch.