Change rr to update patch meta data properly
My intent with this change is that the rr command would properly update
the patch meta data when it has successfully pushed the patch to the
remote.

The reason I am doing this is because I am going to use this state
information to provide more context in the ls command so that the user
would know if they just created the branch, have created a branch and
pushed it up to the remote, have requested review of it, or have
integrated the patch into upstream.

The approach I took for this was to implement a new utility function in
state_management to facilitate the process of reading the metadata from
disk, fetching the meta data for a given patch_id, handing the resulting
Option<Patch> over to a closure that is then responsible for processing
the meta data state change given the previous state and whatever
captured data to produce a new patch meta data object that is then
written back to disk.

Once I had that I simply utilized it in the rr() function to properly
handle updating of patch meta data.