Finish implementing basic list functionality
My intention with this is to finish the basic list functionality so that
it lists the patches in the patch stack, but also includes visual
representations of their states (e.g. integrated, pushed to remote,
branch created, etc.).

I am doing this so that users will actually be able to start using this
version of patch stack.

My approach was to extract the patch id and fetch the associated patch
state for each of the patches in the stack. That way I can compute a
visual representation of the state and show it to the user.