Add show command to the cli
I add the show command to the command line interface so that users can
easily see the raw form of an identified patch.

To do this I added the show() function to the ps module and to the main
library entry point. This exposes the function so that I could then use
it in the porcelain show command in the cli.

I also moved the subcommand about messages to the automatically detected
comments as it reads nicer.