Add sync command to the cli
I am adding the sync command to the command line interface so that users
that want to sync/backup their patches to the remote can easily do so
using the sync command.

THe approach I took was to first build the ps/plumbing/sync module and
sync() function and add it to the ps module. I then created the
commands/plumbing/sync module and function, added it to the
commands/plumbing module and then used it in the main command line
application entry point.