Add patch_index optionality to checkout command
Add optionality of patch_index in the checkout command so that in the
future I will be able to use the lack of a patch_index being present to
trigger a reset operation that will checkout the branch that was checked
out prior to checking out the isolated patch.

The approach I took for this was simply to make the command line options
property an Option<>. structopt then recognizes that and uses it to
proprely generate the help messaging. Beyond that I changed the
patch_index argument of both the command checkout() and the ps lib
checkout() function so that they are both Option<> typed. This allowed
me to then set the groundwork for implementing the eventual handling of
this in the future.