Rename checkout command to isolate
Rename the checkout command to isolate so that we can later add the
checkout command back in the form that it was in the first incarnation
of git-ps where it would check out the commit associated with the
specified patch stack index. Leaving you in a headless state. This
functionality is still wanted by users and naming wise checkout as a
name most closely maps to that functionality.

Therefore this newer functionality that is really about isolating a
patch from the stack for various reasons is now being renamed to isolate
as I think that more accurately represents the action being taken.

The approach that I took to accomplish this was simply to move the code
over from ps/checkout.rs file to the ps/isolate.rs file and wire it up
to the command line interface.