Add classic checkout command
Add the classic version of checkout command where it is simply a
convenience functionality that maps a patch-index to do a checkout of a
commit in the classic Git sense of checking out a commit. I am adding
this because from talking to people that used the previous incarnation
of git-ps they valued this functionality for different use cases than we
are addressing with the "isolate" command.

I had the scafolding of this function already in the library from before
when we had called the "isolate" function "checkout" previously.
Therefore I just filled in that scafolding and plumbed it up to the
command line interface.