Change isolate command's associated branch name
Change the isolate command's associated branch name so that it is more
easily identifiable outside of git patch stack if someone was just
looking in git.

My approach for doing this was first extracting the magic strings out to
a constant and then using that constant in the two locations where the
magic strings were located and then changing the value of that that
constant to be the new isolate branch name.