Add branch override to br, sync, rr, & int
Add branch override to the branch, sync, request-review, and integrate
commands so that if a user is stuck working within a branch naming
scheme because of automation, etc. They can use the -n <branch-name>
option to specify the branch name to use. *Note:* If the patch has
previously been associated with a branch name then the branch name
provided via the -n switch on the command line is ignored.

This closes issue #14.