Add branch cleanup to integrate command
Add branch cleanup to integrate command so that once a patch has been
successfully integrated into upstream the branches used behind the
scenes both on the remote and the local machine are cleaned up so we
don't have just a bunch of excess noise.

The approach I took is simply to add the -k and --keep-branch switches
to the integrate command as a boolean flag and pass that down to the
integrate function where I then conditionally check if I should remove
the branches or not. Additionally I cleaned up some error construction
code in the integrate command while I was there to make things a bit
nicer.

This change addresses issue #15.