Add beginnings of patch content checking for int
Add patch content comparison checking for the integrate command so that
it checks first to make sure if the patch in the patch stack is the same
as the patch in the remote branch. This is useful because then I can
make decisions about this. For example if they are the same continue
with the integrate process. If they aren't error out to the user letting
them know they are out of sync and need to resolve the issue.

This is associated with issue #38.