Add git::commit_diff_patch_id()
Add the git::commit_diff_patch_id() function to facilitate getting the
diff patch id associated with a commit. This will be used to help
identify if changes of the actually patch/diff have happened when the
commit shas are different. This will eventually be used in both the list
command and the integrate command. The list command will use it to
identify state of there being differences between the patch that is
synced and the local one and reporting on it. The integrate command will
use it to check the same thing and abort the integrate process if they
are different.

This relates to issues #38 and #12.