Implement new integrate command with -f option
Implement the new integrate command with the -f option for force. This
is the new concept of the integrate command where it effectively does a
bunch of checks to make sure that the patch is properly synced to the
remote before allowing you to integrate that patch. If however you use
the -f option it skips these checks and in fact doesn't even require
that the patch be synced at all.

This resolves issue #38.