Clean up some warnings
Clean up warnings caused by having branch_name when I am not using it. I
corrected this by prefixing it with an underscore even though I am not
using them. This could probably be cleaned up even further in the future
by replacing the unused _branch_name with _ as I am not using it, or
maybe Rust lets us match without the args.