Add remote_name to state_management tracking
Add remote_name to state_management tracking for all cases that have a
remote branch. I did this because I came to realization that a remote
branch is not uniquely identifiable without it's remote name. This will
also facilitate properly implementing the list command so that we can
find the remote branches in cases where we need to.