Rename `branch` command to `request-review-branch`
Rename the `branch` command to `request-review-branch` so that it is
clearer as to what it does. Also I am planning on adding a new command
called `patch-series-branch` and once there are two branch commands I
thought it would be important to create the clear distinction. I also
removed the abbreviated (short) version of the command `br` for two
reasons. First I couldn't come up with a good abbreviation. Second I
decidied that **plumbing** commands shouldn't get abbreviations as they
aren't intended to be used directly on the command line and are intended
more to facilitate scripting. In the rare case that one wants to use a
**plumbing** command on the command line they should have tab completion
to help with the long command name.

The approach I took was to change the command line interface and leave
the underlying code named based on branch(). I will change that in a
following patch.

[changelog]
changed: `branch` command name to `request-review-branch`
removed: the `branch` commands shortname `br`