Rename commands/branch.rs to commands/request_review_branch.rs
Rename the commands/branch.rs file to commands/request_review_branch.rs
and the function exposed in it from branch() to request_review_branch()
so that the naming matches the recent rename of the command from branch
to request_review_branch. This is important as we will be adding a new
command in the near future called patch-series-branch.

My approach with this was to simply rename the commands layer of the CLI
tool so that it matches the recently renamed command. The ps library
portion still needs to be renamed but that will happen in a different
patch.