Rename ps::branch to ps::request_review_branch
Rename the ps::branch to ps::request_review_branch so that the naming of
the command and it's various library implementations would match. I also
did this in preparation for adding a new command `patch-series-branch`.

I renamed all the related functions and error types within the ps module
and the library. This completes the rename of the `branch` command to
the `request-review-branch` command.

[changelog]
changed: ps::branch() function name to ps::request_review_branch()