Add the `branch` command
Add the `branch` command as a bridge out of the land of patch stack and
patch stack concepts like patches & requested reviews into the classic
world of Git and it's concepts of branches and commits. This is intended
to be a mechanism for people take patches that exist in patch stack and
create a normal git branch containing either a singular patch or a
series of patches so that the user can then operate on that git branch
as they normally would in git land.

The approach I took was to follow suite with all the other commands
where the command line tool has the command but we also provide the
command as a function available in the library that the command line
tool utilizes. That way creation of other tools around these workflow
concepts can be easily done.

This resolves issue #13.

[changelog]
added: branch command as a bridge from patch stack to git concepts