Fix bug where `int -f` fail for fresh patches
There is currently a bug where if you try and force integrate a "fresh
patch", a new patch that doesn't have a ps-id in it's description yet,
it will fail out. It should fail out with the error if it isn't a force
integration. However, if it is a force integration it shouldn't error
out and instead should add a ps id to the patch.

The problem looked to be that I had the verification of the ps-id across
both cases when it really should have only been part of the non-force
code path. Hence, I relocated it to be part of the non-force code path
and allowed the force code path to get the ps-id from the
request_review_branch() which already handles adding of ps-id to fresh
patches.

This relates to issue #40.

[changelog]
fixed: bug `int -f` would incorrectly fail out on patches without a ps-id

ps-id: 46c5834a-5dbb-409c-942b-a5a09b467f70