Fix the singular commit check
Fix a bug where list & integration would failing because it was
incorrectly figuring out the number of commits on the associated branch.
This check is known as the singular commit check.

Remove the incorrectly placed adjustment in the common_ancestor()
function and account for the initial case of a patch branch living
within the patch stack within singular_commit_of_branch() by allowing
for a count of 0 commits as long as the common ancestor and branch shas
match.

[changelog]
fixed: error that was incorrectly being return about incorrect number of commits

ps-id: 21ea0257-2ff4-4f57-b8d3-5af4d06fcb59