Add bash & zsh shell completion generation
Add bash & zsh shell completion generation to the build process so that
everytime we build the app it will generate bash & zsh completion
scripts for the command line tool.

My approach was first to pull the structopt Cli stuff into a separate
file src/cli.rs so that I could include it in the root level build.rs
file as well as use it in the src/main.rs. This makes it so that there
is one source of truth for the command line options and that is in the
src/cli.rs file.

[changelog]
added: bash & zsh completion scription generation to build process