Rename the app name from git-ps to gps
Rename the app name from git-ps to gps because gps has been growing on
me and previously we ran into completion issues with using the "git ps
<command>" form rather than the "git-ps <command>" form. Therefore, I
decided we should just stick with gps as a separate command even though
it extends git. This gives us a shorter command, a lot of us already
have this as an alias and it should make it so completion can work
properly.