Update readme
I did this because there is a hard dependency of hooks out of the box
when installing git-ps-rs. This will hopefully help people set it up
correctly when using ti for the first time.

ps-id: 282C88FB-2874-4088-A1BA-F113AD3CA81D

Updating the rr command to point to hooks

Add information about the hard dependency of hooks and request-review

update Readme

Making the note be specific