Fix pull so it doesn't lose commits incorrectly
Fix the pull command so that it no longer incorrectly loses commits in
the process of rebasing.

Turned out that it was the `git pull --rebase` that we were executing
that was actually losing the commits. I dug through a bunch of docs and
things trying to figure out a way to correct it's behavior but there
didn't seem to be anything. Therefore I changed the implementation of
the pull command to do a fetch and then do a rebase, example
`git rebase --onto origin/main origin/main main`, which seems to have
the correct behavior.

This resolves issue #43.

[changelog]
fixed: bug where pull command was losing commits