Add support for GPG commit signing
Add support for signing commits with a GPG signature so that when the
cherry-picking operations are performed the user doesn't lose their
commits are still signed.

My approach was to extract a create_commit() function to handle
appropriately creating the commit either as a signed commit or as an
unsigned commit. This of course involved implementing those two paths as
well as adding some warning logging for ssh or x509 key formats which we
do not currently support. So dumps a warning and falls back to creating
unsigned commits. Once I had the create_commit() function in place I
could then use it everywhere necessary and update the arguments of the
various functions to pass the reference to the git config in.

This is related to issue #46.

[changelog]
added: support for GPG commit signing