Add isolate_post_checkout hook to isolate command
Add the isolate_post_checkout hook to the isolate command so that after
it has successfully checked out that branch the hook can be used to run
things like build scripts and test suites to make sure that the patch is
actually independent, buildable, and testable.

The approach is basically to try and get hook. If it is found then
execute the hook. If it isn't found then print out a warning on standard
error explaining the siutation and that the hook was skipped.

This relates to issue #50.

[changelog]
added: isolate_post_checkout hook to the isolate command

ps-id: 642db3fc-eaab-4fe8-9433-124354cfc4ae