Integrate from local rr branch
Modify the integrate command to update the local rr branch and integrate
from it rather than the remote one. This should resolve the issue of not
being able to integrate multiple patches in a row by the push failing
due to the branches having diverged.

The other thing that could be done to potentially minimize divergent
branches prevent integration in the future would be to do a fetch at the
beginning of the integration command. However, this could slow down the
integrate command and might not be necessary the majority of the time.
Therefore, I opted to just make this change and see how it feels.

I simply used the request_review_branch() function to properly update
the local request review branch and then modified the push to push from
the local branch instead of the remote branch.

This relates to issue #52.

[changelog]
changed: integrate command to eliminate majority of divergent branch errors

ps-id: 5e5c453c-88de-41d0-be57-44494dbf5b5c