Add fetch to beginning of non-force integrate
Add fetch to the beginning of the non-forced integrate command so that
if someone has pushed a commit to remote request review branch we will
be aware of that state when doing the verification steps of the
integrate. If we didn't do this then there would be potential for use to
publish a patch when the pull request actually has the patch plus some
other commit which is likely not what you would want.

This relates to issue #52.

[changelog]
changed: non-forced integrate command to do a fetch before verification

ps-id: 5758dd41-d887-47f0-ae8a-e4464e050210