Add git::uncommitted_changes_exist() func
Add the git::uncommitted_changes_exist() function to make it so that we
can easily identify if uncommted changes exist in the current working
state of the git repository. I am adding this because I plan to use it
to add checking for uncommitted changes to the isolate command so we can
actually verify that it is going to be isolated.

This relates to #55.

ps-id: e433ba1b-6db4-426b-b0ac-96fc72f2e36b