Add uncommitted changes check to isolate command
Add an uncommited changes check to the beginning of the isolate command
so that we can make sure that the checked out temporary isolation branch
is truely in a state of isolation and doesn't accidentally have any
uncommitted changes or untracked files cary over.

This is related to issue #55.

[changelog]
added: uncommited changes check to start of isolate command

ps-id: 7605fac3-5d17-4d74-bd4d-c548fe9ecb75