Setup basics of config support
My intention with this change is to setup the basics of the config
system with the three different tiers of config files user level,
personal repository, and communal repository and with priority
overriding so that the rest of the application could simply be aware of
the PsConfig interface with defaults applied rather than having to worry
about optionals and spreading defaults through out the application.

My approach was to create basically a layer of DTO (Data Transfer
Objects) that the toml is deserialized into. These DTO have optionality
for the values and the sections as they are optional in the config file.
Then I created a PsConfig layer which has no optionality and I created a
`get_config()` function to facilitate reading the 3 layers of configs,
merging them together in priority order and then applying defaults.

This relates to issue #37.