Make pull command explicitly drop dup patches
Make pull command explicitly drop dup patches. It was doing it
previously using the the `git rebase` default behavior. But this now
appears to print a warning informing the user about the
--reapply-cherry-picks option. So we decided to explicitly declare that
we wanted it to drop dup commits by adding the --no-reapply-cherry-picks
option to the rebase command. Hopefully, this will also eliminate the
potentially confusing warning git rebase is printing. If not it still
seems like a good change.

[changelog]
changed: pull command to explicitly drop dup commits when rebasing