Add configurable isolation verification
Add configurable isolation verification to the request-review and
integrate commands so that users can get feedback on the isolation,
buildability, and testability before requesting review or integrating.
If however the user doesn't want this to happen it is configurable via
the config files.

This relates to issue #50.

[changelog]
added: configurable isolation verification to request review & integrate commands

ps-id: e1a5470d-1731-476c-b38c-d6269600b20b