Add ability to config pull cmd to also list patches
Add ability to configure the pull command to list patches after it has
completed fetching and rebasing. This is disabled by default. But it can
be enabled in any of the layers of config files as follows.

  [pull]
  show_list_post_pull = true

The approach I took was simply to read the config and at the end if the
configuration value is enabled then call the list command.

This relates to issue #8.

[changelog]
added: configuration to enable listing patches at end of the pull command

ps-id: 58ab021d-bf4a-492f-81a1-da28ae33605f