Add configurable user verification to integrate cmd
Add a configurable user verification step to the integrate command so
that when enabled it will present people with details about the patch
and then prompt them to verify they actually want to integrate the
patch. By default this feature is enabled but can be controlled within
any of the config file layers with the following.

  [integrate]
  require_verification = false

The approach I took was simply adding the check for the state of the
config value within the non-force path. If enabled then I have it do the
show command for the patch and then prompt for the user's verification
before deciding to bail out or proceed.

This is related to issue #27.

ps-id: 228c2b18-d43a-40cc-aa8c-2245a0235e39