Fix bug where iso wasn't cleaning up its branch
The isolate command was not cleaning up its temporary branch after the
user was resetting back to the patch stack. This was happening because
we simply forgot the code to make it clean up it's branch. This change
corrects this by adding the missing code.

This is related to issue #56.

[changelog]
fixed: bug where the isolate command wasn't cleaning up it's branch

ps-id: 440fa8d9-8e92-4d80-83cf-d2f871cbde4f