Add create-patch (c) command to the cli
Add the create-patch (c) command to the cli so that users can easily
create patches from stagged changes without having to break the
abstraction of Git Patch Stack and drop down to the Git layer of
concepts. Also this helps us enforce best practices like not allowing
users to use the `git commit -m` and making sure they are using the `git
commit -v` switch.

This relates to issue #24.

[changelog]
added: create-patch (c) command to the cli

ps-id: 9677e862-6bbe-4f18-b5f1-fefc6c8b473d