Add amend-patch command to the cli
Add the amend-patch command to the cli so that people can discover it
and know that they can amend the top most patch of the patch stack. This
also helps to hold the Git Patch Stack abstraction so that people don't
have to drop down to Git and think about Git concepts.

This relates to issue #62.

[changelog]
added: amend-patch (a) command to the command line interface

ps-id: 03e065c9-584e-4da8-b00d-30056270651c