Add the --continue switch to the rebase command
Add the --continue switch to the rebase command so that users don't have
to leave the Git Patch Stack abstraction to continue a rebase that was
started via Git Patch Stack.

I simply added the --continue options to the command line interface as a
boolean and then threaded the boolean for continue_rebase through the
command interface and then into the ps::public::rebase() function.

This relates to issue #67.

[changelog]
added: --continue switch to the rebase command

ps-id: b47566e9-0c93-401a-8baa-dfe1c6130089